requirements.txt
openai==1.39.0
python-dotenv==1.0.0
requests==2.31.0
Pillow==11.0.0
instagrapi==2.1.3
schedule==1.2.2
gunicorn==20.1.0
uvicorn==0.27.0
fastapi==0.109.0
pytest==8.3.3
pyyaml==6.0.1
